Almost completely finished with the front end
Polished everything up. The whole front end should basically work, just
be sure to only click on the first result for things like courses and
users, since the rest are dummy data and won't build the page. I think
we basically just need to hook up the end points. And of course test it
a bit. The Instructor toggle should work for seeing what an instructor
would see. Logging out will clear the user, but a refresh will reset the
user. It's still hardcoded for now.